+++ lib/gem-licenses/install_tasks.rb
@@ -1,0 +1,10 @@
+module Gem
+  module Licenses
+    def self.install_tasks
+      spec = Gem::Specification.find_by_name('gem-licenses')
+      Dir[File.join(spec.gem_dir, 'tasks/*.rake')].each do |file|
+        load file
+      end
+    end
+  end
+end