+++ lib/gem_licenses.rb
@@ -1,3 +1,4 @@
 require 'gem-licenses/version'
+require 'gem-licenses/install_tasks'
 require 'gem/specification'
-require 'gem/gem'
+require 'gem/licenses'