+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 #### Next
 
+* Added `gem:licenses` Rake task - [@dblock](https://github.com/dblock).
+* Install Rake tasks via `Gem::Licenses.install_tasks` - [@dblock](https://github.com/dblock).
 * Your contribution here.
 
 #### 0.1.2 (7/27/2011)